+++ src/pop/op/longitudinal/n_mode.cc
@@ -21,7 +21,7 @@ namespace FT{
             /// Evaluates the node and updates the state states. 
             void NodeMode::evaluate(const Data& data, State& state)
             {
-                ArrayXd tmp(state.z.top().first.size());
+                ArrayXf tmp(state.z.top().first.size());
                 
                 int x;
                 
@@ -30,7 +30,7 @@ namespace FT{
                     
                 state.z.pop();
 
-                state.push<double>(tmp);
+                state.push<float>(tmp);
                 
             }
             #else
@@ -51,7 +51,7 @@ namespace FT{
             /// Evaluates the node symbolically
             void NodeMode::eval_eqn(State& state)
             {
-                state.push<double>("mean(" + state.zs.pop() + ")");
+                state.push<float>("mean(" + state.zs.pop() + ")");
             }
             
             NodeMode* NodeMode::clone_impl() const { return new NodeMode(*this); }
