+++ src/pop/op/longitudinal/n_skew.cc
@@ -22,7 +22,7 @@ namespace FT{
             /// Evaluates the node and updates the state states. 
             void NodeSkew::evaluate(const Data& data, State& state)
             {
-                ArrayXd tmp(state.z.top().first.size());
+                ArrayXf tmp(state.z.top().first.size());
                 
                 int x;
                 
@@ -31,7 +31,7 @@ namespace FT{
                     
                 state.z.pop();
 
-                state.push<double>(tmp);
+                state.push<float>(tmp);
                 
             }
             #else
@@ -52,7 +52,7 @@ namespace FT{
             /// Evaluates the node symbolically
             void NodeSkew::eval_eqn(State& state)
             {
-                state.push<double>("skew(" + state.zs.pop() + ")");
+                state.push<float>("skew(" + state.zs.pop() + ")");
             }
 
             NodeSkew* NodeSkew::clone_impl() const { return new NodeSkew(*this); }
